Crate pallet_connectors

Re-exports

pub use pallet::*;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

weights

Autogenerated weights for pallet_connectors

Structs

DomainAddress

DomainLocator

XcmDomain

XcmDomain gathers all the required fields to build and send remote calls to a specific XCM-based Domain.

Enums

Domain

A Domain is a chain or network we can send a Connectors message to. The domain indices need to match those used in the EVM contracts and these need to pass the Centrifuge domain to send tranche tokens from the other domain here. Therefore, DO NOT remove or move variants around.

Message

ParachainId

The Parachains that Centrifuge Connectors support.

Router

Constants

TOKEN_NAME_SIZE

The fixed size for the array representing a tranche token name

TOKEN_SYMBOL_SIZE

Functions

encoded_contract_call

Return the encoded contract call, i.e, ConnectorsXcmRouter::handle(encoded_msg).

xcm_router_contract

The ConnectorsXcmRouter Abi as in ethabi::Contract Note: We only concern ourselves with the handle function of the contract since that’s all we need to build the calls for remote EVM execution.

Type Definitions

CurrencyIdOf

MessageOf

PoolIdOf

TrancheIdOf